nexus3/pkg/repository/apt_hosted.go
package repository

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/datadrivers/go-nexus-client/nexus3/pkg/client"
	"github.com/datadrivers/go-nexus-client/nexus3/pkg/tools"
	"github.com/datadrivers/go-nexus-client/nexus3/schema/repository"
)

const (
	aptHostedAPIEndpoint = repositoryAPIEndpoint + "/apt/hosted"
)

type RepositoryAptHostedService struct {
	client *client.Client
}

func NewRepositoryAptHostedService(c *client.Client) *RepositoryAptHostedService {
	return &RepositoryAptHostedService{
		client: c,
	}
}

func (s *RepositoryAptHostedService) Create(repo repository.AptHostedRepository) error {
	data, err := tools.JsonMarshalInterfaceToIOReader(repo)
	if err != nil {
		return err
	}
	body, resp, err := s.client.Post(aptHostedAPIEndpoint, data)
	if err != nil {
		return err
	}
	if resp.StatusCode != http.StatusOK && resp.StatusCode != http.StatusCreated {
		return fmt.Errorf("could not create repository '%s': HTTP: %d, %s", repo.Name, resp.StatusCode, string(body))
	}
	return nil
}

func (s *RepositoryAptHostedService) Get(id string) (*repository.AptHostedRepository, error) {
	var repo repository.AptHostedRepository
	body, resp, err := s.client.Get(fmt.Sprintf("%s/%s", aptHostedAPIEndpoint, id), nil)
	if err != nil {
		return nil, err
	}
	if resp.StatusCode != http.StatusOK {
		return nil, fmt.Errorf("could not read repository '%s': HTTP: %d, %s", id, resp.StatusCode, string(body))
	}
	if err := json.Unmarshal(body, &repo); err != nil {
		return nil, fmt.Errorf("could not unmarshal repository: %v", err)
	}
	return &repo, nil
}

func (s *RepositoryAptHostedService) Update(id string, repo repository.AptHostedRepository) error {
	data, err := tools.JsonMarshalInterfaceToIOReader(repo)
	if err != nil {
		return err
	}
	body, resp, err := s.client.Put(fmt.Sprintf("%s/%s", aptHostedAPIEndpoint, id), data)
	if err != nil {
		return err
	}
	if resp.StatusCode != http.StatusNoContent {
		return fmt.Errorf("could not update repository '%s': HTTP: %d, %s", id, resp.StatusCode, string(body))
	}
	return nil
}

func (s *RepositoryAptHostedService) Delete(id string) error {
	return deleteRepository(s.client, id)
}
